UnitCategory convert: Call UnitCategoryPrivate instead of duplicating it
This is probably a mistake when implementing the private class. Both
UnitCategoryPrivate::convert and UnitCategory::convert essentially do
the same thing. However, all sub categories modify the
UnitCategoryPrivate virtual method, so we should be calling that one.

This was caught while trying to debug the currency converter. It has its
own custom convert function.

REVIEW: 122181