Make currency conversion table fetching explicit
By converting the first currency unit value inside a process, a blocking
and automatic network sync of the currency conversion table is done. In
order to give the framework user more control, introduce API for
explicitly triggering this sync at a convenience point in the startup
process. Any later currency unit value creation will then skip a sync
during a grace period of 24 hours (which also is the previous behavior).

For KF6, it is a good idea to fully remove the blocking call, but that
is a behavior change.