Add conversions between fabric weights
These are common units of fabric weight for clothing, but intuitive
familiarity with only one scale or the other can leave people confused
and make rash purchase decisions when first encountering the other one.

(This may or may not have happened to me, I'm admitting nothing!)

This commit adds conversions between them to make comparison and
understanding easier. I opted to add a new category for them as
they're kind of halfway between a weight and a density.